requirements.txt
Django==3.1
djangorestframework==3.11.1
psycopg2==2.8.5
Pillow==7.2.0

pypng==0.0.20
PyQRCode==1.2.1
pyzbar==0.1.8
qrcode==6.1

flake8==3.8.3